+++ src/core/backend/avx512/fft/mod.rs
@@ -1,0 +1,188 @@
+use std::arch::x86_64::{
+    __m512i, _mm512_add_epi32, _mm512_broadcast_i64x4, _mm512_load_epi32, _mm512_min_epu32,
+    _mm512_permutexvar_epi32, _mm512_store_epi32, _mm512_sub_epi32, _mm512_xor_epi32,
+};
+
+pub mod ifft;
+pub mod rfft;
+
+/// An input to _mm512_permutex2var_epi32, and is used to interleave the even words of a
+/// with the even words of b.
+const EVENS_INTERLEAVE_EVENS: __m512i = unsafe {
+    core::mem::transmute([
+        0b00000, 0b10000, 0b00010, 0b10010, 0b00100, 0b10100, 0b00110, 0b10110, 0b01000, 0b11000,
+        0b01010, 0b11010, 0b01100, 0b11100, 0b01110, 0b11110,
+    ])
+};
+/// An input to _mm512_permutex2var_epi32, and is used to interleave the odd words of a
+/// with the odd words of b.
+const ODDS_INTERLEAVE_ODDS: __m512i = unsafe {
+    core::mem::transmute([
+        0b00001, 0b10001, 0b00011, 0b10011, 0b00101, 0b10101, 0b00111, 0b10111, 0b01001, 0b11001,
+        0b01011, 0b11011, 0b01101, 0b11101, 0b01111, 0b11111,
+    ])
+};
+
+/// An input to _mm512_permutex2var_epi32, and is used to concat the even words of a
+/// with the even words of b.
+const EVENS_CONCAT_EVENS: __m512i = unsafe {
+    core::mem::transmute([
+        0b00000, 0b00010, 0b00100, 0b00110, 0b01000, 0b01010, 0b01100, 0b01110, 0b10000, 0b10010,
+        0b10100, 0b10110, 0b11000, 0b11010, 0b11100, 0b11110,
+    ])
+};
+/// An input to _mm512_permutex2var_epi32, and is used to concat the odd words of a
+/// with the odd words of b.
+const ODDS_CONCAT_ODDS: __m512i = unsafe {
+    core::mem::transmute([
+        0b00001, 0b00011, 0b00101, 0b00111, 0b01001, 0b01011, 0b01101, 0b01111, 0b10001, 0b10011,
+        0b10101, 0b10111, 0b11001, 0b11011, 0b11101, 0b11111,
+    ])
+};
+/// An input to _mm512_permutex2var_epi32, and is used to interleave the low half of a
+/// with the low half of b.
+const LHALF_INTERLEAVE_LHALF: __m512i = unsafe {
+    core::mem::transmute([
+        0b00000, 0b10000, 0b00001, 0b10001, 0b00010, 0b10010, 0b00011, 0b10011, 0b00100, 0b10100,
+        0b00101, 0b10101, 0b00110, 0b10110, 0b00111, 0b10111,
+    ])
+};
+/// An input to _mm512_permutex2var_epi32, and is used to interleave the high half of a
+/// with the high half of b.
+const HHALF_INTERLEAVE_HHALF: __m512i = unsafe {
+    core::mem::transmute([
+        0b01000, 0b11000, 0b01001, 0b11001, 0b01010, 0b11010, 0b01011, 0b11011, 0b01100, 0b11100,
+        0b01101, 0b11101, 0b01110, 0b11110, 0b01111, 0b11111,
+    ])
+};
+const P: __m512i = unsafe { core::mem::transmute([(1u32 << 31) - 1; 16]) };
+
+// TODO(spapini): FFTs return a redundant representation, that can get the value P. need to reduce
+// it somewhere.
+
+/// Transposes the AVX vectors in the given array.
+/// Swaps the bit index abc <-> cba, where |a|=|c| and |b| = 0 or 1, according to the parity of
+/// `log_n_vecs`.
+/// When log_n_vecs is odd, transforms the index abc <-> cba, w
+///
+/// # Safety
+/// This function is unsafe because it takes a raw pointer to i32 values.
+/// `values` must be aligned to 64 bytes.
+///
+/// # Arguments
+/// * `values`: A mutable pointer to the values that are to be transposed.
+/// * `log_n_vecs`: The log of the number of AVX vectors in the `values` array.
+pub unsafe fn transpose_vecs(values: *mut i32, log_n_vecs: usize) {
+    let half = log_n_vecs / 2;
+    for b in 0..(1 << (log_n_vecs & 1)) {
+        for a in 0..(1 << half) {
+            for c in 0..(1 << half) {
+                let i = (a << (log_n_vecs - half)) | (b << half) | c;
+                let j = (c << (log_n_vecs - half)) | (b << half) | a;
+                if i >= j {
+                    continue;
+                }
+                let val0 = _mm512_load_epi32(values.add(i << 4).cast_const());
+                let val1 = _mm512_load_epi32(values.add(j << 4).cast_const());
+                _mm512_store_epi32(values.add(i << 4), val1);
+                _mm512_store_epi32(values.add(j << 4), val0);
+            }
+        }
+    }
+}
+
+/// Computes the twiddles for the first fft layer from the second, and loads both to AVX registers.
+/// Returns the twiddles for the first layer and the twiddles for the second layer.
+/// # Safety
+unsafe fn compute_first_twiddles(twiddle1_dbl: [i32; 8]) -> (__m512i, __m512i) {
+    // Start by loading the twiddles for the second layer (layer 1):
+    // The twiddles for layer 1 are replicated in the following pattern:
+    //   0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7
+    let t1 = _mm512_broadcast_i64x4(std::mem::transmute(twiddle1_dbl));
+
+    // The twiddles for layer 0 can be computed from the twiddles for layer 1:
+    // A circle coset of size 4 in bit reversed order looks like this:
+    //   [(x, y), (-x, -y), (y, -x), (-y, x)]
+    // Note: This is related to the choice of M31_CIRCLE_GEN, and the fact the a quarter rotation
+    //   is (0,-1) and not (0,1). This would cause another relation.
+    // The twiddles for layer 0 are the y coordinates:
+    //   [y, -y, -x, x]
+    // The twiddles for layer 1 in bit reversed order are the x coordinates:
+    //   [x, y]
+    // Works also for inverse of the twiddles.
+
+    // The twiddles for layer 0 are computed like this:
+    //   t0[4i:4i+3] = [t1[2i+1], -t1[2i+1], -t1[2i], t1[2i]]
+    const INDICES_FROM_T1: __m512i = unsafe {
+        core::mem::transmute([
+            0b0001, 0b0001, 0b0000, 0b0000, 0b0011, 0b0011, 0b0010, 0b0010, 0b0101, 0b0101, 0b0100,
+            0b0100, 0b0111, 0b0111, 0b0110, 0b0110,
+        ])
+    };
+    // Xoring a double twiddle with 2^32-2 transforms it to the double of it negation.
+    // Note that this keeps the values as a double of a value in the range [0, P].
+    const NEGATION_MASK: __m512i = unsafe {
+        core::mem::transmute([0i32, -2, -2, 0, 0, -2, -2, 0, 0, -2, -2, 0, 0, -2, -2, 0])
+    };
+    let t0 = _mm512_xor_epi32(_mm512_permutexvar_epi32(INDICES_FROM_T1, t1), NEGATION_MASK);
+    (t0, t1)
+}
+
+// TODO(spapini): Move these to M31 AVX.
+
+/// Adds two packed M31 elements, and reduces the result to the range [0,P].
+/// Each value is assumed to be in unreduced form, [0, P] including P.
+/// # Safety
+/// This function is safe.
+pub unsafe fn add_mod_p(a: __m512i, b: __m512i) -> __m512i {
+    // Add word by word. Each word is in the range [0, 2P].
+    let c = _mm512_add_epi32(a, b);
+    // Apply min(c, c-P) to each word.
+    // When c in [P,2P], then c-P in [0,P] which is always less than [P,2P].
+    // When c in [0,P-1], then c-P in [2^32-P,2^32-1] which is always greater than [0,P-1].
+    _mm512_min_epu32(c, _mm512_sub_epi32(c, P))
+}
+
+/// Subtracts two packed M31 elements, and reduces the result to the range [0,P].
+/// Each value is assumed to be in unreduced form, [0, P] including P.
+/// # Safety
+/// This function is safe.
+pub unsafe fn sub_mod_p(a: __m512i, b: __m512i) -> __m512i {
+    // Subtract word by word. Each word is in the range [-P, P].
+    let c = _mm512_sub_epi32(a, b);
+    // Apply min(c, c+P) to each word.
+    // When c in [0,P], then c+P in [P,2P] which is always greater than [0,P].
+    // When c in [2^32-P,2^32-1], then c+P in [0,P-1] which is always less than [2^32-P,2^32-1].
+    _mm512_min_epu32(_mm512_add_epi32(c, P), c)
+}
+
+#[cfg(all(target_arch = "x86_64", target_feature = "avx512f"))]
+#[cfg(test)]
+mod tests {
+    use self::ifft::get_itwiddle_dbls;
+    use super::*;
+    use crate::core::fields::m31::BaseField;
+    use crate::core::poly::circle::CanonicCoset;
+
+    #[test]
+    fn test_twiddle_relation() {
+        let ts = get_itwiddle_dbls(CanonicCoset::new(5).circle_domain());
+        let t0 = ts[0]
+            .iter()
+            .copied()
+            .map(|x| BaseField::from_u32_unchecked((x as u32) / 2))
+            .collect::<Vec<_>>();
+        let t1 = ts[1]
+            .iter()
+            .copied()
+            .map(|x| BaseField::from_u32_unchecked((x as u32) / 2))
+            .collect::<Vec<_>>();
+
+        for i in 0..t0.len() / 4 {
+            assert_eq!(t0[i * 4], t1[i * 2 + 1]);
+            assert_eq!(t0[i * 4 + 1], -t1[i * 2 + 1]);
+            assert_eq!(t0[i * 4 + 2], -t1[i * 2]);
+            assert_eq!(t0[i * 4 + 3], t1[i * 2]);
+        }
+    }
+}
